feat(worker): determine source column based on non null fields
Previously, the source field would be converted to 'custom' if the variant came from a variant ID manually input. This meant that when re-processing a list, all the source columns would turn to Custom. Now source column is determined by the presence of a value for the 'lof' field, for gnomAD, and the 'gold_stars' field, for ClinVar